requirements.txt
### dependency
tqdm
dvc
pandas
numpy
PyYAML
tensorflow
tensorflow_datasets
mkdocs-material
joblib
notebook
### local packages -
-e . # << uncomment this to install from setup.py
